+++ crates/perl-ast/src/classification.rs
@@ -1140,6 +1140,7 @@ mod tests {
             NodeKind::Subroutine {
                 name: Some("foo".to_string()),
                 name_span: None,
+                declarator: None,
                 prototype: None,
                 signature: None,
                 attributes: vec![],
@@ -1272,6 +1273,7 @@ mod tests {
             NodeKind::Subroutine {
                 name: Some("foo".to_string()),
                 name_span: None,
+                declarator: None,
                 prototype: None,
                 signature: None,
                 attributes: vec![],
@@ -1642,6 +1644,7 @@ mod tests {
             n(NodeKind::Subroutine {
                 name: Some("foo".to_string()),
                 name_span: None,
+                declarator: None,
                 prototype: Some(Box::new(Node::new(
                     NodeKind::Prototype { content: "$@".to_string() },
                     loc(),